add models

+++ src/main/java/com/suleimanov/vehiclecontrol/vehicle/models/VehicleModel.java
@@ -22,6 +22,5 @@ public class VehicleModel extends CommonObject {
   @JoinColumn(name = "vehiclemakeid", insertable = false, updatable = false)
   @JsonIgnore
   private VehicleMake vehiclemake;
-
-  private Integer vehiclemakeid;
+  private Long vehiclemakeid;
 }
